refactor: 💡 Redesign AppHeader component ui
